requirements.txt
Flask==3.0.0
Pillow==10.1.0
boto3==1.28.0
python-dotenv==1.0.0
werkzeug==3.0.1
pycognito